Make InSim.add_button's text argument mandatory
8a4fe2b1 changed the text parameter to Variant to allow mixing strings
and Callables for text mapping across connections, but it introduced
an error in GdUnit tests, which seems to be confused by the default
value (empty string) when mocking the class, and apparently typing this
parameter as a String instead, causing a mismatch with the actual
function signature (see #728 on the GdUnit4 GitHub repository).

This commit makes the text parameter mandatory; this shouldn't be an
issue for usability, as empty strings aren't that common anyway.