Use second IS_STA packet to decrease risk of wrong packet count
If a player leaves after LFS sends the IS_STA packet, but before LFS
receives our TINY_NPL request (unlikely but technically possible),
initialization is stuck because of a missing packet.

Using a second IS_STA limits this risk by decreasing the time interval
that allows the issue to occur.